Completed Sprints and Some Theory
Due to working on my personal device, I have not committed or pushed any of the progress that I have made in the past few weeks.

Firstly, I have added the gun texture and even found a tutorial that would help me include a knife, allowing for easy weapon switching. These were added around three weeks ago, approximately the 30th of May this year. I have also removed the image of the crosshair from the previous commit as I could not get it to work.

Secondly, I improved the bullet's positioning and added enemies in the 'second' sprint. Since I had already included a bit of Object Oriented Programming into my project, I was told I was actually up to the third sprint rather than the second. From here I ended the second sprint and moved onto the third. I needed to improve my enemy mechanics; they followed me, but they did not take damage when the bullet hit them.

Thirdly, I mentioned that I had found a tutorial by Ursina Tutorials on YouTube that allowed the player to switch weapons. This was quite object oriented, including the classes for the player and the bullet. I modified these classes to increase functionality and changed the enemy class as the enemies would just go to the player's spawn point and stop. I also finally added enemy hit mechanics and rounds, which will be changed later (around 10/6). During this process whenever the bullet hit a wall  or the ground the game would crash due to not having an attribute "health" so I created a new Map class for them and set their health to infinite.

Finally, today I completed all the theory for sprint 1 (I am quite behind on the theory). In today's next commit I plan to have completed at least half of the sprint 2 and 3 theory so that I can get up to speed.

Message for future commits: The plan is to add some more functionality to the game such as a boss round, the necessary splash screens and enemy health bars. Also need to make the enemies solid so that they do not merge with one another or the player. This will improve the functionality of the game.