Finished Sprint 3 Build & Sprint 2 Theory
Unfortunately, yesterday I did not finish sprint 3 as planned. This meant that I didn't commit as I said that I would, however, today, that has finally been completed (the sprint 3 theory is yet to come). In the next commit, I hope to have all theory up to sprint 4 well done and polished, ready to submit. This will put me on track for the final sprint, and a completed and fully functional program for submission.