Completed Sprint 3 and Most of Sprint 4
In this commit I completed all of sprint 3 and most of sprint 4. This means that I have just the final tweaking of the code, the launch and evaluation that I need to do before the task is complete. These will be done in either the next commit or the next two commits.