Completed Everything Except for The Peer Evaluation
In this commit I have completed every bit of the theory and practical except for the peer evaluation. This puts me in a tough position as the project is due at 9:00 AM today, and the current time is around 1:35 AM. I will just have to get two of my peers to quickly evaluate my project before the deadline.